third_party/blink/web_tests/platform/generic/editing/inserting/insert-empty-html-expected.txt

This test ensures inserting empty string does not move the caret.

The caret is placed at 5th pos and the empty html is added. The test passes if the caret is positioned at the same position after the execution insertion command.

abcdefghijklmnopqrstuvwxyz
PASS window.getSelection().baseOffset is 5
PASS successfullyParsed is true

TEST COMPLETE
